Update LSF Finetune (#189)
Revise ```MoiraiFinetune```, dataset creatrion and associated config
files for LSF tasks. Provide scripts for finetuning
```moirai-1.0-small``` on ETTh and ETTm datasets as examples. Use the
multivariate setup for ETTh and the univariate setup for ETTm. The Mixin
has not been applied to ```MoiraiFinetune``` yet. This can be further
discussed and set as a to-do task. Code links will be added to the
README after the PR is merged.

Co-authored-by: Chenghao Liu <74166079+chenghaoliu89@users.noreply.github.com>